% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LOD}
\alias{LOD}
\title{Compute the LOD score of a marker given a data set}
\usage{
LOD(marker, data, phenotype = "IgG1")
}
\arguments{
\item{marker}{Marker name (e.g. "Marker_1", "Marker_2", should be a column of \code{data}).}

\item{data}{Data frame with genotypes and phenotype. Genotypes should be coded 0/1}

\item{phenotype}{Phenotype name (defaults to "IgG1"), should be a column of \code{data}).}
}
\value{
The LOD score (numeric value)
}
\description{
Compute the LOD score of a marker given a data set
}
\examples{
data(hw_data)
LOD('Marker_1', data = hw_data)
}
