% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LOD_position}
\alias{LOD_position}
\title{Compute the LOD score at a given position (unobserved) given a data set and a genetic map}
\usage{
LOD_position(position, data, map, phenotype = "IgG1")
}
\arguments{
\item{position}{Position (in cM) of the marker along the chromosome}

\item{data}{Data frame with genotypes and phenotype. Genotypes should be coded 0/1}

\item{phenotype}{Phenotype name (defaults to "IgG1"), should be a column of \code{data}).}

\item{genetic_map}{Genetic map of the chromosome}
}
\value{
LOD score (numeric value)
}
\description{
Compute the LOD score at a given position (unobserved) given a data set and a genetic map
}
\examples{
data(hw_data)
data(genetic_map)
LOD_position(position = 48, data = hw_data, map = genetic_map)
}
