% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reconstruction}
\alias{reconstruction}
\title{Reconstruct allele based on flanking markers}
\usage{
reconstruction(la, ra, ld, rd)
}
\arguments{
\item{la}{Allele (0/1) at left flanking marker}

\item{ra}{Allele (0/1) at right flanking marker}

\item{ld}{Distance (in Morgans) to left flanking marker}

\item{rd}{Distance (in Morgans) to right flanking marker}
}
\value{
Probability of allele 0 at given position
}
\description{
Reconstruct allele based on flanking markers
}
\examples{
## close to marker with allele 0, likely to be 0
reconstruction(la = 0, ra = 1, ld = 0.02, rd = 0.18)
## close to marker with allele 1, unlikely to be 0
reconstruction(la = 0, ra = 1, ld = 0.18, rd = 0.02)
}
